+++ RELEASE_NOTES.md
@@ -1,5 +1,16 @@
 # PyGPSClient Release Notes
 
+### RELEASE 1.4.20
+
+FIXES:
+
+1. Fixes typo in ubx_handler which affects NAV2-STATUS messages [#142](https://github.com/semuconsulting/PyGPSClient/issues/142)
+
+CHANGES:
+
+1. Minor improvements to settings frame appearance.
+1. Add support for logging in underlying `pygnssutils` utilities (`gnssntripclient`, `gnssmqttclient`, `gnssserver`)
+
 ### RELEASE 1.4.19
 
 CHANGES: